Feat: SSE logs, dashboard UI, fixes SQL/XML/SignalR
## API (.NET)
- PricePusher: remplace `dynamic` par `SessionDto` typé — élimine le RuntimeBinderException
- StationService: cast `::float8` sur price/distanceKm — résout l'incompatibilité Dapper decimal↔double
- LogBroadcaster + LogBroadcastProvider: stream SSE des logs .NET sur GET /logs
- Program.cs: enregistrement LogBroadcaster + endpoint /logs SSE

## Ingestion (Go)
- parser.go: CharsetReader ISO-8859-1 sans dépendance externe — résout l'erreur de décodage XML
- main.go: logHub SSE + broadcastWriter — stream des logs ingestion sur :5001/logs

## Frontend
- Dashboard layout : sidebar (prix héro + meta) + main (top 5 cliquable)
- Top 5 : chaque ligne est sélectionnable, met à jour le héro et le bouton Y aller
- Bouton "Y aller →" : itinéraire Google Maps depuis la position utilisateur
- Géoloc fallback : autocomplete villes via api-adresse.data.gouv.fr si GPS refusé
- Boot : validation de session côté serveur avant affichage — retour accueil si session invalide
- SignalR : guard `typeof signalR` → fallback polling si CDN indisponible

## Infra
- docker-compose: expose :5001 (ingestion SSE), API remappée sur 5050
- scripts/dev.sh: trap SIGINT/SIGTERM → docker compose down propre + open frontend auto
- scripts/test_api.go: test CLI coloré de tous les endpoints (health, meta, nearby, setup, cheapest, SSE)

+++ api/Services/StationService.cs
@@ -25,11 +25,11 @@ public class StationService
                 gs.city          AS City,
                 gs.postal_code   AS PostalCode,
                 fp.fuel_type     AS FuelType,
-                fp.price         AS Price,
+                fp.price::float8 AS Price,
                 ROUND(
-                    ST_Distance(gs.location, ST_SetSRID(ST_MakePoint(@lng, @lat), 4326)::geography)
-                    / 1000.0, 2
-                )                AS DistanceKm
+                    (ST_Distance(gs.location, ST_SetSRID(ST_MakePoint(@lng, @lat), 4326)::geography)
+                    / 1000.0)::numeric, 2
+                )::float8        AS DistanceKm
             FROM fuel_prices fp
             JOIN gas_stations gs ON gs.id = fp.station_id
             WHERE
@@ -65,11 +65,11 @@ public class StationService
                 gs.city          AS City,
                 gs.postal_code   AS PostalCode,
                 fp.fuel_type     AS FuelType,
-                fp.price         AS Price,
+                fp.price::float8 AS Price,
                 ROUND(
-                    ST_Distance(gs.location, ST_SetSRID(ST_MakePoint(@lng, @lat), 4326)::geography)
-                    / 1000.0, 2
-                )                AS DistanceKm
+                    (ST_Distance(gs.location, ST_SetSRID(ST_MakePoint(@lng, @lat), 4326)::geography)
+                    / 1000.0)::numeric, 2
+                )::float8        AS DistanceKm
             FROM fuel_prices fp
             JOIN gas_stations gs ON gs.id = fp.station_id
             WHERE
